requirements.txt
aiodns==1.2.0
aiomysql
aiohttp==3.5.4
aiohttp-cors==0.7.0
async-timeout==3.0.1
attrs==18.2.0
cchardet==2.1.4
chardet==3.0.4
idna==2.8
multidict==4.5.2
peewee-async==0.5.12
peewee-moves==1.7.3
pycares==2.4.0
typing==3.6.6
yarl==1.3.0
